% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{cles2d}
\alias{cles2d}
\title{Common language effect size to Cohen's D}
\usage{
cles2d(cles)
}
\arguments{
\item{cles}{Common language effect size}
}
\value{
Cohen's D value
}
\description{
Common language effect size to Cohen's D
}
\examples{
cles2d(0.8)

}
\references{
http://rpsychologist.com/d3/cohend/
}
