% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{cohend_convert}
\alias{cohend_convert}
\title{Cohen's D Effect Size Descriptions}
\usage{
cohend_convert(d = NULL, overlap = NULL, u3 = NULL, cles = NULL)
}
\arguments{
\item{d}{Cohen's D. Optional.}

\item{overlap}{Distributional overlap. Optional.}

\item{u3}{Cohen's U3. Optional.}

\item{cles}{Common language effect size. Optional.}
}
\value{
List containing all the Cohen's D descriptors
}
\description{
Convert between different descriptions of Cohen's D. Enter only one
descriptor, and obtain all the others.
}
\examples{
cohend_convert(d=0.8)
cohend_convert(u3=0.5)

}
\references{
http://rpsychologist.com/d3/cohend/
}
