% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{d2cles}
\alias{d2cles}
\title{Convert Cohen's D to Common Language Effect Size (CLES)}
\usage{
d2cles(d)
}
\arguments{
\item{d}{Cohen's D}
}
\value{
The common language effect size (CLES)
}
\description{
Calculate the common language effect size from a Cohen's D: the probability
that an individual from each group selected at random will be higher or lower
in the expected direction than the individual from the other group.
}
\examples{
d2cles(0.8)

}
\references{
http://rpsychologist.com/d3/cohend/
}
