% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{d2n_crit}
\alias{d2n_crit}
\title{Calculate the number of measurements for which a Cohen's D is significant}
\usage{
d2n_crit(d, sig.level = 0.05, alternative = "two.sided")
}
\arguments{
\item{d}{Cohen's D}

\item{sig.level}{Significance level. Optional: default is 0.05.}

\item{alternative}{Alternative hypothesis. Optional: default is "two.sided"}
}
\value{
The number of measurements
}
\description{
There's almost definitely a better way of doing this.
}
\examples{
d2n_crit(0.5)

}
