% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{d2n_crit_model}
\alias{d2n_crit_model}
\title{Model for calculating the number of measurements for which a Cohen's D is significant}
\usage{
d2n_crit_model(n, d, sig.level = 0.05)
}
\arguments{
\item{n}{Number of measurements}

\item{d}{Cohen's D}

\item{sig.level}{Significance level. Optional: default is 0.05.}
}
\value{
The difference from the significance level.
}
\description{
There's almost definitely a better way of doing this.
}
\examples{
d2n_crit_model(50, 0.5)

}
