% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{d2n_powersummary}
\alias{d2n_powersummary}
\title{Power analysis summary for a given Cohen's D value}
\usage{
d2n_powersummary(d, sig.level = 0.05, alternative = "two.sided")
}
\arguments{
\item{d}{Cohen's D}

\item{sig.level}{Significance level. Optional: default is 0.05}

\item{alternative}{Alternative hypothesis. Optional: default is "two.sided"}
}
\value{
A data.frame with the sample sizes for different levels of power.
}
\description{
Power analysis summary for a given Cohen's D value
}
\examples{
d2n_powersummary(d=0.8)

}
