% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{d2overlap}
\alias{d2overlap}
\title{Convert Cohen's D to overlap}
\usage{
d2overlap(d)
}
\arguments{
\item{d}{Cohen's D}
}
\value{
The percentage overlap of the group distributions
}
\description{
Calculate the distributional overlap percentage for a given Cohen's D effect size.
}
\examples{
d2overlap(0.8)

}
\references{
http://rpsychologist.com/d3/cohend/
}
