% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{d2u3}
\alias{d2u3}
\title{Convert Cohen's D to Cohen's U3}
\usage{
d2u3(d)
}
\arguments{
\item{d}{Cohen's D}
}
\value{
Cohen's U3
}
\description{
Calculate the percentage of each distribution which will not overlap with the
mean of the other distribution
}
\examples{


}
\references{
http://rpsychologist.com/d3/cohend/
}
