% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{d_crit}
\alias{d_crit}
\title{Calculate the critical value of Cohen's D}
\usage{
d_crit(n1, n2 = n1, sig.level = 0.05, alternative = "two.sided")
}
\arguments{
\item{n1}{Number of measurements in group 1}

\item{n2}{Number of measurements in group 2. Optional: default is n1.}

\item{sig.level}{Significance level. Optional: default is 0.05}

\item{alternative}{Alternative hypothesis. Optional: default is "two.sided"}
}
\value{
The critical Cohen's D value
}
\description{
Calculate the critical value of D for which a given sample size will yield a significant result
}
\examples{
d_crit(20)

}
