% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{d_nunnally}
\alias{d_nunnally}
\title{Calculate the Cohen's D after attenuation based on Nunnally's equation}
\usage{
d_nunnally(rel_total = NULL, rel_onegroup = NULL, d = NULL,
  overlap = NULL, u3 = NULL, cles = NULL, n1 = 10, n2 = n1)
}
\arguments{
\item{rel_total}{Reliability or reliabilities for the total combined sample of measurements}

\item{rel_onegroup}{Reliability or reliabilities for one group of measurements.}

\item{d}{True Cohen's D of biological signal before accounting for measurement error}

\item{overlap}{True distributional overlap of biological signal before accounting for measurement error}

\item{u3}{True Cohen's U3 of biological signal before accounting for measurement error}

\item{cles}{True CLES of biological signal before accounting for measurement error}

\item{n1}{Number of measurements in group 1. Optional: only required if group sizes differ.}

\item{n2}{Number of measurements in group 2. Optional: only required if group sizes differ.}
}
\value{
The Cohen's D after attenuation
}
\description{
Enter in only one effect size measure
}
\examples{
d_nunnally(rel_total = 0.7, u3=0.9)
d_nunnally(rel_onegroup = 0.7, u3=0.9)
}
