% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{dmax_attenuation}
\alias{dmax_attenuation}
\title{Calculate the maximum Cohen's D based on Spearman's equation}
\usage{
dmax_attenuation(...)
}
\arguments{
\item{...}{A vector of reliability scores of the measures included}
}
\value{
The maximum Cohen's D value for a given reliability
}
\description{
Calculate the maximum Cohen's D based on Spearman's equation
}
\examples{
dmax_attenuation(0.8, 0.7)
}
