% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{dmax_nunnally}
\alias{dmax_nunnally}
\title{Calculate the maximum Cohen's D based on Nunnally's equation}
\usage{
dmax_nunnally(...)
}
\arguments{
\item{...}{A vector of reliability scores of the measures included}
}
\value{
The maximum Cohen's D value for a given reliability
}
\description{
Calculate the maximum Cohen's D based on Nunnally's equation
}
\examples{
dmax_nunnally(0.8, 0.7)
}
