% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{extrapRel2sd}
\alias{extrapRel2sd}
\title{Extrapolate reliability to standard deviation}
\usage{
extrapRel2sd(icc_extrap, icc_original, sd_original = 1, rho = 1)
}
\arguments{
\item{icc_extrap}{Extrapolated ICC value (i.e. desired reliability)}

\item{icc_original}{Original ICC value}

\item{sd_original}{Original SD value. Optional: default is 1. For 1, it will be the proportional increase in the SD.}

\item{rho}{SEM inflation factor. Optional: default is 1. For 1, the SEM is assumed to be constant between studies.}
}
\value{
The standard deviation of a new sample for which the ICC would be equal to the desired value
}
\description{
Calculate the standard deviation required to extrapolate an ICC value to a desired ICC value
}
\examples{
extrapRel2sd(icc_extrap=0.9, icc_original=0.5)

}
