% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{icc2sem}
\alias{icc2sem}
\title{Convert ICC values to SEM values}
\usage{
icc2sem(icc, sd)
}
\arguments{
\item{icc}{ICC value}

\item{sd}{Standard deviation of the whole sample}
}
\value{
The SEM (standard error of measurement) value
}
\description{
Convert ICC values to SEM values
}
\examples{
icc2sem(icc=0.9, sd=2)

}
