% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{mean_tot}
\alias{mean_tot}
\title{Calculate the total mean}
\usage{
mean_tot(n1, mean1, n2 = n1, mean2)
}
\arguments{
\item{n1}{Number of measurements in group 1.}

\item{mean1}{Mean of group 1.}

\item{n2}{Number of measurements in group 2.}

\item{mean2}{Mean of group 2.}
}
\value{
The total mean.
}
\description{
Calculates the total mean over two groups.
}
\examples{
mean_tot(n1=15, mean1=5, n2=10, mean2=3)

}
