% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{n2d_powersummary}
\alias{n2d_powersummary}
\title{Power analysis summary of Cohen's D values for a given sample size}
\usage{
n2d_powersummary(n1, n2 = n1, sig.level = 0.05, alternative = "two.sided")
}
\arguments{
\item{n1}{Number of measurements in group 1.}

\item{n2}{Number of measurements in group 2. Optional: defaults to n1.}

\item{sig.level}{Significance level. Optional: default is 0.05}

\item{alternative}{Alternative hypothesis. Optional: default is "two.sided"}
}
\value{
A data.frame with the Cohen's D effect sizes for different levels of power.
}
\description{
Power analysis summary of Cohen's D values for a given sample size
}
\examples{
n2d_powersummary(n1=20)
}
