% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{n2r_powersummary}
\alias{n2r_powersummary}
\title{Power analysis summary of Pearson's r values for a given sample size}
\usage{
n2r_powersummary(n, sig.level = 0.05, alternative = "two.sided")
}
\arguments{
\item{n}{Number of measurements}

\item{sig.level}{Significance level. Optional: default is 0.05}

\item{alternative}{Alternative hypothesis. Optional: default is "two.sided"}
}
\value{
A data.frame with the effect sizes for a given sample size.
}
\description{
Power analysis summary of Pearson's r values for a given sample size
}
\examples{
n2r_powersummary(50)

}
