% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_difference}
\alias{plot_difference}
\title{Plot the difference between two distributions}
\usage{
plot_difference(d = NULL, overlap = NULL, u3 = NULL, cles = NULL,
  colours = c("#85d4e3", "#e39f85"))
}
\arguments{
\item{d}{Cohen's D. Optional: pick only one effect size}

\item{overlap}{Distributional overlap.  Optional: pick only one effect size}

\item{u3}{Cohen's U3.  Optional: pick only one effect size}

\item{cles}{Common Language Effect Size.  Optional: pick only one effect size}

\item{colours}{Colours of the two distributions. Optional.}
}
\value{
A ggplot2 object of the plot
}
\description{
Visualise the difference between two distributions given an effect size measure
}
\examples{
plot_difference(d=2)

}
\references{
http://rpsychologist.com/d3/cohend/
}
