% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{r2n_crit_model}
\alias{r2n_crit_model}
\title{Model for calculating the number of measurements for which a Pearson's r will be significant}
\usage{
r2n_crit_model(n, r, sig.level = 0.05)
}
\arguments{
\item{n}{Number of measurements}

\item{r}{Pearson's r: correlation coefficient}

\item{sig.level}{Significance level. Optional: default is 0.05.}
}
\value{
The difference from the significance level.
}
\description{
Model for calculating the number of measurements for which a Pearson's r will be significant
}
\examples{
r2n_crit_model(50, 0.5)

}
