% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{r2n_powersummary}
\alias{r2n_powersummary}
\title{Power analysis summary for a given Pearson's r value}
\usage{
r2n_powersummary(r, sig.level = 0.05, alternative = "two.sided")
}
\arguments{
\item{r}{Pearson's r: correlation coefficient}

\item{sig.level}{Significance level. Optional: default is 0.05}

\item{alternative}{Alternative hypothesis. Optional: default is "two.sided"}
}
\value{
A data.frame with the power sample sizes for a given effect size.
}
\description{
Power analysis summary for a given Pearson's r value
}
\examples{
r2n_powersummary(0.5)

}
