% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{r_attenuation}
\alias{r_attenuation}
\title{Attenuation factor for correlation coefficients based on Spearman's equation}
\usage{
r_attenuation(...)
}
\arguments{
\item{...}{A vector of reliability scores of the measures included}
}
\value{
The degree to which a correlation coefficient is attenuated by the measurement error
}
\description{
Attenuation factor for correlation coefficients based on Spearman's equation
}
\examples{
r_attenuation(0.8, 0.7, 0.7)

}
