% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{r_crit}
\alias{r_crit}
\title{Calculate the critical value of Pearson's r}
\usage{
r_crit(n, sig.level = 0.05, alternative = "two.sided")
}
\arguments{
\item{n}{Number of measurements-}

\item{sig.level}{Significance level. Optional: default is 0.05}

\item{alternative}{Alternative hypothesis. Optional: default is "two.sided"}
}
\value{
The critical r value
}
\description{
Calculate the critical value of r for which a given sample size will yield a significant result
}
\examples{
r_crit(50)

}
