% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{r_nunnally}
\alias{r_nunnally}
\title{Attenuation factor for correlation coefficients based on Nunnally's equation}
\usage{
r_nunnally(...)
}
\arguments{
\item{...}{A vector of reliability scores of the measures included}
}
\value{
The degree to which a correlation coefficient is attenuated by the measurement error
}
\description{
Attenuation factor for correlation coefficients based on Nunnally's equation
}
\examples{
r_nunnally(0.8, 0.7, 0.7)

}
