% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relfeas-package.R
\docType{package}
\name{relfeas-package}
\alias{relfeas}
\alias{relfeas-package}
\title{relfeas: Reliability for Feasibility Analysis}
\description{
This package contains functions for examining the reliability of outcome measures for use in assessment of the feasibility of planned studies. It contains functions by which to calculate the extrapolated reliability, to calculate signal attenuation, to plot effect sizes to to visually judge their size, as well as convenience functions to convert quantities.
}
\author{
\strong{Maintainer}: Granville Matheson \email{mathesong@gmail.com}

}
\keyword{internal}
