% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{roundbound}
\alias{roundbound}
\title{Bounded rounding}
\usage{
roundbound(x, digits = 0, maximum = 100, minimum = 0)
}
\arguments{
\item{x}{a numeric value.}

\item{digits}{integer indicating the number of decimal places}

\item{maximum}{maximum bound. Defaults to 100.}

\item{minimum}{minimum bound. Defaults to 0.}
}
\value{
A character string of the rounded value
}
\description{
This function is for presenting a figure when it should not reach a certain
bound, and instead shows it as greater than the number below.
}
\examples{
roundbound(99.9)
roundbound(99.3)

roundbound(0.9)
roundbound(0.3)

}
