% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{sd2extrapRel}
\alias{sd2extrapRel}
\title{Extrapolate standard deviation to reliability}
\usage{
sd2extrapRel(sd, icc_original, sd_original = 1, tau = 1)
}
\arguments{
\item{sd}{Standard deviation in the new sample for which the reliability is being extrapolated.}

\item{icc_original}{ICC from the original study.}

\item{sd_original}{Standard deviation from the original study. Optional: default is 1. For 1, the sd argument is the proportional increase in SD.}

\item{tau}{SEM inflation factor. Optional: default is 1. For 1, the SEM is assumed to be constant between studies.}
}
\value{
The extrapolated ICC ICC value for the new sample
}
\description{
Calculate the extrapolated reliability for a given standard deviation value
}
\examples{
sd2extrapRel(sd=2, icc_original=0.5)

}
