% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{sd_pooled}
\alias{sd_pooled}
\title{Calculate the pooled standard deviation}
\usage{
sd_pooled(n1, sd1, n2 = n1, sd2 = sd1)
}
\arguments{
\item{n1}{Number of measurements in group 1.}

\item{sd1}{SD of group 1.}

\item{n2}{Number of measurements in group 2. Optional: default is n1.}

\item{sd2}{SD of group 2. Optional: default is sd1.}
}
\value{
The pooled standard deviation
}
\description{
Calculates the pooled standard deviation from summary statistics. This is not
the total standard deviation over groups, but the pooled standard deviation
used for calculation of the Cohen's D.
}
\examples{
sd_pooled(n1=10, sd1=3, n2=15, sd2=5)

}
