% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{sd_tot}
\alias{sd_tot}
\title{Calculate the total standard deviation}
\usage{
sd_tot(n1, mean1, sd1, n2 = n1, mean2 = NULL, sd2 = sd1, d = NULL)
}
\arguments{
\item{n1}{Number of measurements in group 1.}

\item{mean1}{Mean of group 1.}

\item{sd1}{SD of group 1.}

\item{n2}{Number of measurements in group 2. Optional: defaults to n1.}

\item{mean2}{Mean of group 2. Optional: this or the effect size can be specified.}

\item{sd2}{SD of group 2. Optional: defaults to sd1.}

\item{d}{Cohen's D. Optional: this or mean2 can be specified.}
}
\value{
Total standard deviation across two groups.
}
\description{
Calculates the total standard deviation across two groups as if they were one combined group.
}
\examples{
sd_tot(n1=20, mean1=1, sd1=1, mean2=5)

}
