% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{sdtot2mean2}
\alias{sdtot2mean2}
\title{Calculate mean2 from the total SD}
\usage{
sdtot2mean2(sd_total, n1, mean1 = 0, sd1 = 1, n2 = n1, sd2 = sd1)
}
\arguments{
\item{sd_total}{Total SD over groups: i.e. of the combined sample.}

\item{n1}{Number of measurements in group 1.}

\item{mean1}{Mean of group 1. Optional: defaults to 0 if only interested in Cohen's D output.}

\item{sd1}{SD of group 1. Optional: default is 1, in which case sd_total is a proportional increase in SD.}

\item{n2}{Number of measurements in group 2. Optional: defaults to n1.}

\item{sd2}{SD of group 2. Optional: defaults to sd1.}
}
\value{
List containing mean2 in positive and negative directions, the total
mean in positive and negative directions, Cohen's D and the percentage
difference.
}
\description{
Calculate mean2 from the total SD
}
\examples{
sdtot2mean2(sd_total=1.2, n1=20)

}
