% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{sem2icc}
\alias{sem2icc}
\title{Convert SEM values to ICC values}
\usage{
sem2icc(sem, sd)
}
\arguments{
\item{sem}{SEM (standard error of measurement) value}

\item{sd}{standard deviation}
}
\value{
The ICC value
}
\description{
Convert SEM values to ICC values
}
\examples{
sem2icc(sem=0.63, sd=2)

}
