% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{sumStat_total}
\alias{sumStat_total}
\title{Summary statistics of combined samples}
\usage{
sumStat_total(n1, mean1, sd1, n2 = n1, mean2 = NULL, sd2 = sd1,
  d = NULL, effect = "increase")
}
\arguments{
\item{n1}{Number of measurements in group 1}

\item{mean1}{Mean of group 1}

\item{sd1}{SD of group 1}

\item{n2}{Number of measurements in group 2. Optional: default is n1.}

\item{mean2}{Mean of group 2. Optional: this or the effect size can be specified.}

\item{sd2}{SD of group 2}

\item{d}{Cohen's D. Optional: this or mean2 can be specified.}

\item{effect}{Direction of the effect. "increase" or "decrease".}
}
\value{
List containing the total n, mean and SD as well as the pooled SD, Cohen's D and mean2.
}
\description{
Calculate the summary statistics over the combined groups from the individual summary statistics
}
\examples{
sumStat_total(n1=20, mean1=5, sd1=1, d=0.8)

}
