% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trt.R
\name{trt_widify}
\alias{trt_widify}
\title{Long to wide format of test-retest data}
\usage{
trt_widify(data, values, cases, rater = NULL)
}
\arguments{
\item{data}{The total data frame}

\item{values}{Column of the measured values (quoted character)}

\item{cases}{Column of the individual values, i.e. individuals/measured items
(character)}

\item{rater}{Optional. Column of the rater/which measurement, i.e.
measurement 1 and measurement, or raters 1, 2 and 3. If this is not
specified, the function will assume that the values are in order by rater,
or that it the order doesn't matter.}
}
\value{
A tibble containing the test-retest data in wide format.
}
\description{
Convenience function to turn long-format test-retest data to wide format.
}
\examples{
library(tidyverse)
data("petVT", package = 'agRee')
amy <- as.data.frame(petVT$amygdala)
amy_l <- gather(amy) \%>\%
  mutate(indiv = rep(paste0('s', 1:11), times=2))

trt_widify(amy_l, 'value', 'indiv', 'key')


}
