% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{u32d}
\alias{u32d}
\title{Convert Cohen's U3 to Cohen's D}
\usage{
u32d(u3)
}
\arguments{
\item{u3}{Cohen's U3 (between 0 and 1)}
}
\value{
Cohen's D
}
\description{
Calculate the Cohen's D for a Cohen's U3: the percentage of each distribution
which will not overlap with the mean of the other distribution
}
\examples{
u32d(0.4)

}
\references{
http://rpsychologist.com/d3/cohend/
}
